MemberList from DB

+++ docker/db/sql/0001-create-member.sql
@@ -4,7 +4,7 @@ DROP TABLE IF EXISTS `member`;
 ---- create ----
 CREATE TABLE member
 (
-    id int NOT NULL AUTO_INCREMENT,
+    id bigint NOT NULL AUTO_INCREMENT,
     name varchar(32) NOT NULL,
     PRIMARY KEY (id)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
